+++ framework/model/src/main/java/is/codion/framework/model/ForeignKeyConditionModel.java
@@ -55,12 +55,16 @@ public interface ForeignKeyConditionModel extends ConditionModel<Entity> {
 	interface Builder {
 
 		/**
+		 * Note that this search model is linked to the EQUAL operand, so no
+		 * linking is required when constructing a UI component.
 		 * @param equalSearchModel the search model to use for the EQUAl operator
 		 * @return this builder
 		 */
 		Builder equalSearchModel(EntitySearchModel equalSearchModel);
 
 		/**
+		 * Note that this search model is linked to the IN operand, so no
+		 * linking is required when constructing a UI component.
 		 * @param inSearchModel the search model to use for the IN operator
 		 * @return this builder
 		 */